Change test vector for missing-0 tables
When omitting 0 from the initial table (ie., writing [x=1..2^k-1] instead of [x=0..2^k-1],
it is impossible to tell whether entries are empty (used in BucketStore and sorting), as
previously they were stored as (f1(x), x) instead of (x,).  For performance, it is better
to omit 0 from the table than to keep track of it throughout.
As a result, a few matches that use (x=0) will not make it into the final f7 table, and
the final plot file will change.